+++ logger/doc.go
@@ -1,6 +1,6 @@
 // Tideland Go Library - Logger
 //
-// Copyright (C) 2012-2015 Frank Mueller / Tideland / Oldenburg / Germany
+// Copyright (C) 2012-2016 Frank Mueller / Tideland / Oldenburg / Germany
 //
 // All rights reserved. Use of this source code is governed
 // by the new BSD license.
@@ -21,9 +21,9 @@ import (
 // VERSION
 //--------------------
 
-// PackageVersion returns the version of the version package.
-func PackageVersion() version.Version {
-	return version.New(4, 2, 0)
+// Version returns the version of the logger package.
+func Version() version.Version {
+	return version.New(4, 3, 0)
 }
 
 // EOF